+++ internal/runtime/runtime.go
@@ -1,24 +1,18 @@
 package runtime
 
 import (
-	"bufio"
 	"context"
 	"crypto/rand"
 	"encoding/hex"
 	"errors"
 	"fmt"
 	"io"
-	"maps"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"slices"
-	"sort"
 	"strconv"
 	"strings"
-	"sync"
 	"time"
-	"unicode"
 	"unicode/utf8"
 
 	"github.com/buildkite/buildkite-gha/internal/action/metadata"
@@ -29,22 +23,13 @@ import (
 const (
 	defaultCleanupTimeout    = 10 * time.Second
 	defaultPostActionTimeout = 10 * time.Minute
-	// Match actions/runner ProcessInvoker's graceful cancellation windows.
-	defaultInterruptGrace = 7500 * time.Millisecond
-	defaultTerminateGrace = 2500 * time.Millisecond
-	maxStreamLineBytes    = 1024 * 1024
 	// Match Buildkite's maximum annotation body size.
 	maxJobAnnotationBytes         = 1024 * 1024
 	jobSummaryAnnotationHeading   = "<h2 class=\"h4 mb2\">GitHub Actions job summary</h2>\n"
 	jobSummaryAnnotationSeparator = "<div class=\"border-top border-gray pt2\"></div>\n\n"
 	maxJobSummaryBytes            = maxJobAnnotationBytes - len(jobSummaryAnnotationHeading) - len(jobSummaryAnnotationSeparator)
 
-	jobSummaryTruncationNotice       = "\n\n---\n_Job summary truncated at the 1 MiB limit._\n"
-	workflowCommandTruncationNotice  = "\n\n---\n_Workflow command annotations truncated at the 1 MiB limit._\n"
-	workflowWarningAnnotationHeading = "<h2 class=\"h4 mb2\">GitHub Actions warnings</h2>\n"
-	workflowErrorAnnotationHeading   = "<h2 class=\"h4 mb2\">GitHub Actions errors</h2>\n"
-	workflowCommandListHeading       = "<div class=\"mb2\">\n"
-	workflowCommandListEnd           = "</div>\n"
+	jobSummaryTruncationNotice = "\n\n---\n_Job summary truncated at the 1 MiB limit._\n"
 )
 
 // Runner executes verified actions using explicitly configured host tools.
@@ -777,228 +762,6 @@ func resultContains(result Result, value string) bool {
 	return false
 }
 
-func (r *jobRun) runProcess(ctx context.Context, processor *commandProcessor, dir string, env map[string]string, result *Result, state map[string]string, name string, args ...string) error {
-	var files commandFiles
-	var err error
-	if r.jobContainer != nil {
-		files, err = newCommandFilesUnder(r.runnerTemp)
-	} else {
-		files, err = newCommandFiles()
-	}
-	if err != nil {
-		return err
-	}
-	defer func() { _ = files.cleanup() }()
-	commandPaths := map[string]string{"GITHUB_OUTPUT": files.output, "GITHUB_ENV": files.env, "GITHUB_PATH": files.path, "GITHUB_STATE": files.state, "GITHUB_STEP_SUMMARY": files.summary}
-	if r.jobContainer != nil {
-		if err := files.allowContainerWrites(); err != nil {
-			return err
-		}
-		for key, path := range commandPaths {
-			commandPaths[key] = r.jobContainer.containerPath(path)
-		}
-	}
-	env = mergeStringMaps(env, commandPaths)
-	var runErr error
-	if r.jobContainer != nil {
-		runErr = r.jobContainer.exec(ctx, r.Runner, processor, dir, env, name, args...)
-	} else {
-		runErr = r.runStreaming(ctx, processor, dir, env, name, args...)
-	}
-	runErr = markStepProcessExit(runErr)
-	effects, fileErr := files.apply(result, state)
-	effects.reportSummaryUploadFailure(processor)
-	if fileErr == nil && (effects.pathSet || len(effects.paths) > 0) {
-		pathEnv := map[string]string{"PATH": env["PATH"]}
-		if effects.pathSet {
-			pathEnv["PATH"] = effects.pathBase
-		}
-		applyPaths(pathEnv, effects.paths)
-		result.Env["PATH"] = pathEnv["PATH"]
-	}
-	return errors.Join(runErr, fileErr)
-}
-
-func (effects fileCommandEffects) reportSummaryUploadFailure(processor *commandProcessor) {
-	if effects.summaryBytes <= maxCommandFileBytes {
-		return
-	}
-	_ = processor.process(processor.stderr, fmt.Sprintf("GITHUB_STEP_SUMMARY upload skipped: content is %d bytes; maximum is %d bytes", effects.summaryBytes, maxCommandFileBytes))
-}
-
-func (r Runner) runStreaming(ctx context.Context, processor *commandProcessor, dir string, env map[string]string, name string, args ...string) error {
-	if err := validateEnvironmentNames(env); err != nil {
-		return err
-	}
-	cmd := exec.Command(name, args...)
-	cmd.Dir = dir
-	cmd.Env = processEnv(env)
-	return r.runStreamingCommand(ctx, processor, cmd)
-}
-
-func resolveExecutableInPath(name, path string) (string, error) {
-	if strings.ContainsRune(name, filepath.Separator) {
-		return name, nil
-	}
-	for _, dir := range filepath.SplitList(path) {
-		if dir == "" {
-			dir = "."
-		}
-		candidate := filepath.Join(dir, name)
-		info, err := os.Stat(candidate)
-		if err == nil && !info.IsDir() && info.Mode().Perm()&0o111 != 0 {
-			return candidate, nil
-		}
-	}
-	return "", &exec.Error{Name: name, Err: exec.ErrNotFound}
-}
-
-func (r Runner) runStreamingCommand(ctx context.Context, processor *commandProcessor, cmd *exec.Cmd) error {
-	configureProcessGroup(cmd)
-	stdout, stdoutWriter, err := os.Pipe()
-	if err != nil {
-		return err
-	}
-	stderr, stderrWriter, err := os.Pipe()
-	if err != nil {
-		_ = stdout.Close()
-		_ = stdoutWriter.Close()
-		return err
-	}
-	cmd.Stdout = stdoutWriter
-	cmd.Stderr = stderrWriter
-	if err := cmd.Start(); err != nil {
-		_ = stdout.Close()
-		_ = stdoutWriter.Close()
-		_ = stderr.Close()
-		_ = stderrWriter.Close()
-		return err
-	}
-	_ = stdoutWriter.Close()
-	_ = stderrWriter.Close()
-	defer func() {
-		_ = stdout.Close()
-		_ = stderr.Close()
-	}()
-
-	processDone := make(chan struct{})
-	var waitErr error
-	go func() {
-		waitErr = cmd.Wait()
-		close(processDone)
-	}()
-	finished := make(chan struct{})
-	terminationDone := make(chan struct{})
-	go func() {
-		defer close(terminationDone)
-		terminateProcessGroup(ctx, cmd.Process.Pid, r.interruptGrace(), r.terminateGrace(), finished)
-	}()
-
-	var wg sync.WaitGroup
-	streamErrs := make([]error, 2)
-	stream := func(index int, label string, reader io.Reader, target io.Writer) {
-		defer wg.Done()
-		var commandErr error
-		streamErr := streamLines(reader, func(line string) {
-			commandErr = errors.Join(commandErr, processor.process(target, line))
-		}, processor.suppress)
-		if streamErr != nil {
-			streamErr = fmt.Errorf("%s stream: %w", label, streamErr)
-		}
-		if commandErr != nil {
-			commandErr = fmt.Errorf("%s stream: %w", label, commandErr)
-		}
-		streamErrs[index] = errors.Join(streamErr, commandErr)
-	}
-	wg.Add(2)
-	go stream(0, "stdout", stdout, processor.stdout)
-	go stream(1, "stderr", stderr, processor.stderr)
-	wg.Wait()
-	<-processDone
-	close(finished)
-	<-terminationDone
-	if ctx.Err() != nil {
-		waitErr = ctx.Err()
-	}
-	if waitErr != nil {
-		waitErr = fmt.Errorf("process %s: %w", cmd.Path, waitErr)
-	}
-	return errors.Join(waitErr, streamErrs[0], streamErrs[1])
-}
-
-func (r Runner) interruptGrace() time.Duration {
-	if r.InterruptGrace > 0 {
-		return r.InterruptGrace
-	}
-	return defaultInterruptGrace
-}
-
-func (r Runner) terminateGrace() time.Duration {
-	if r.TerminateGrace > 0 {
-		return r.TerminateGrace
-	}
-	return defaultTerminateGrace
-}
-
-func streamLines(reader io.Reader, process func(string), suppress func()) error {
-	buffered := bufio.NewReader(reader)
-	line := make([]byte, 0, buffered.Size())
-	oversized := false
-	sawOversized := false
-	for {
-		fragment, err := buffered.ReadSlice('\n')
-		if !oversized {
-			contentBytes := len(line) + len(fragment)
-			if err == nil {
-				contentBytes--
-				if len(fragment) > 1 && fragment[len(fragment)-2] == '\r' {
-					contentBytes--
-				}
-			}
-			if contentBytes > maxStreamLineBytes {
-				line = line[:0]
-				oversized = true
-				sawOversized = true
-				suppress()
-			} else {
-				line = append(line, fragment...)
-			}
-		}
-
-		if err == nil {
-			if oversized {
-			} else {
-				process(trimStreamLineEnding(string(line)))
-			}
-			line = line[:0]
-			oversized = false
-			continue
-		}
-		if errors.Is(err, bufio.ErrBufferFull) {
-			continue
-		}
-		if errors.Is(err, io.EOF) {
-			if oversized {
-			} else if len(line) != 0 {
-				process(trimStreamLineEnding(string(line)))
-			}
-			if sawOversized {
-				return fmt.Errorf("line exceeds %d-byte limit and was discarded", maxStreamLineBytes)
-			}
-			return nil
-		}
-		if sawOversized {
-			return errors.Join(fmt.Errorf("line exceeds %d-byte limit and was discarded", maxStreamLineBytes), err)
-		}
-		return err
-	}
-}
-
-func trimStreamLineEnding(line string) string {
-	line = strings.TrimSuffix(line, "\n")
-	return strings.TrimSuffix(line, "\r")
-}
-
 func newResult() Result {
 	return Result{Outputs: make(map[string]string), Env: make(map[string]string), State: make(map[string]string)}
 }
@@ -1024,370 +787,4 @@ func actionInputEnv(inputs map[string]string) map[string]string {
 		env["INPUT_"+name] = value
 	}
 	return env
-}
-
-func mapEnv(values map[string]string) []string {
-	env := make([]string, 0, len(values))
-	for _, name := range sortedKeys(values) {
-		env = append(env, name+"="+values[name])
-	}
-	return env
-}
-
-func processEnv(overrides map[string]string) []string {
-	// This allowlist is also the mise trust boundary: ambient MISE_* values must
-	// never redirect compatibility runtime downloads or verification.
-	values := make(map[string]string, 6+len(overrides))
-	for _, name := range []string{"PATH", "HOME", "TMPDIR", "LANG", "LC_ALL", "LC_CTYPE"} {
-		if value, ok := os.LookupEnv(name); ok {
-			values[name] = value
-		}
-	}
-	if _, ok := values["PATH"]; !ok {
-		values["PATH"] = "/usr/local/bin:/usr/bin:/bin"
-	}
-	if _, ok := values["HOME"]; !ok {
-		if home, err := os.UserHomeDir(); err == nil {
-			values["HOME"] = home
-		}
-	}
-	if _, ok := values["TMPDIR"]; !ok {
-		values["TMPDIR"] = os.TempDir()
-	}
-	maps.Copy(values, overrides)
-	return mapEnv(values)
-}
-
-func validateEnvironmentNames(env map[string]string) error {
-	for name := range env {
-		if name == "" || strings.ContainsAny(name, "=\x00") {
-			return fmt.Errorf("invalid environment variable name %q", name)
-		}
-	}
-	return nil
-}
-
-func sortedKeys[V any](values map[string]V) []string {
-	keys := make([]string, 0, len(values))
-	for key := range values {
-		keys = append(keys, key)
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-type commandProcessor struct {
-	mu              sync.Mutex
-	stdout          io.Writer
-	stderr          io.Writer
-	masks           []string
-	trustedWarnings workflowCommandAnnotationBuffer
-	warnings        workflowCommandAnnotationBuffer
-	errors          workflowCommandAnnotationBuffer
-	stopToken       string
-	discard         bool
-}
-
-type workflowCommandAnnotationBuffer struct {
-	commands  []workflowCommandAnnotation
-	rendered  int
-	truncated bool
-}
-
-type workflowCommandAnnotation struct {
-	file     string
-	title    string
-	location string
-	message  string
-}
-
-type parsedWorkflowCommand struct {
-	name       string
-	properties map[string]string
-	message    string
-}
-
-func newCommandProcessor(stdout, stderr io.Writer) *commandProcessor {
-	return &commandProcessor{stdout: stdout, stderr: stderr}
-}
-
-var errInvalidWorkflowCommandStopToken = errors.New("invalid ::stop-commands workflow command")
-
-func (p *commandProcessor) process(target io.Writer, line string) error {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	if p.discard {
-		return nil
-	}
-	command, isCommand := parseWorkflowCommand(line)
-	if p.stopToken != "" {
-		if isCommand && strings.EqualFold(command.name, p.stopToken) {
-			p.stopToken = ""
-		}
-		p.writeMaskedLineLocked(target, line)
-		return nil
-	}
-	if isCommand {
-		switch {
-		case strings.EqualFold(command.name, "add-mask"):
-			p.addMaskLocked(command.message)
-			return nil
-		case strings.EqualFold(command.name, "stop-commands"):
-			if !validWorkflowCommandStopToken(command.message) {
-				const message = "invalid ::stop-commands token: token is empty or collides with a workflow command"
-				p.appendWorkflowCommandLocked(&p.errors, workflowErrorAnnotationHeading, parsedWorkflowCommand{message: message})
-				p.writeWorkflowCommandMessageLocked(target, "error", message)
-				return errInvalidWorkflowCommandStopToken
-			}
-			p.stopToken = command.message
-			if len(command.message) > 6 {
-				p.addMaskLocked(command.message)
-			}
-			p.writeMaskedLineLocked(target, line)
-			return nil
-		case strings.EqualFold(command.name, "warning"):
-			p.appendWorkflowCommandLocked(&p.warnings, workflowWarningAnnotationHeading, command)
-			p.writeWorkflowCommandMessageLocked(target, "warning", command.message)
-			return nil
-		case strings.EqualFold(command.name, "error"):
-			p.appendWorkflowCommandLocked(&p.errors, workflowErrorAnnotationHeading, command)
-			p.writeWorkflowCommandMessageLocked(target, "error", command.message)
-			return nil
-		case strings.EqualFold(command.name, "group"):
-			p.writeLogSectionLocked(target, command.message)
-			return nil
-		case strings.EqualFold(command.name, "endgroup"),
-			strings.EqualFold(command.name, "debug"),
-			strings.EqualFold(command.name, "add-matcher"),
-			strings.EqualFold(command.name, "remove-matcher"):
-			return nil
-		}
-	}
-	if isLegacyPresentationCommand(line) {
-		return nil
-	}
-	p.writeMaskedLineLocked(target, line)
-	return nil
-}
-
-func (p *commandProcessor) logSection(name string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.writeLogSectionLocked(p.stdout, name)
-}
-
-func (p *commandProcessor) expandCurrentSection() {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	_, _ = fmt.Fprintln(p.stdout, "^^^ +++")
-}
-
-func (p *commandProcessor) writeLogSectionLocked(target io.Writer, name string) {
-	name = sanitizeLogSectionText(p.maskTextLocked(name))
-	if name != "" {
-		_, _ = fmt.Fprintln(target, "--- "+name)
-	}
-}
-
-func sanitizeLogSectionText(text string) string {
-	text = strings.Map(func(r rune) rune {
-		if r == '\r' || r == '\n' || unicode.IsControl(r) {
-			return ' '
-		}
-		return r
-	}, text)
-	return strings.TrimSpace(text)
-}
-
-func isLegacyPresentationCommand(line string) bool {
-	line = strings.TrimLeftFunc(line, unicode.IsSpace)
-	for _, command := range []string{"##[debug]", "##[add-matcher]", "##[remove-matcher]"} {
-		if strings.HasPrefix(strings.ToLower(line), command) {
-			return true
-		}
-	}
-	return false
-}
-
-func validWorkflowCommandStopToken(token string) bool {
-	if token == "" || strings.EqualFold(token, "pause-logging") {
-		return false
-	}
-	for _, command := range []string{
-		"add-mask", "add-matcher", "add-path", "debug", "echo", "endgroup", "error", "group",
-		"internal-set-repo-path", "notice", "remove-matcher", "save-state", "set-env",
-		"set-output", "set-repo-path", "stop-commands", "warning",
-	} {
-		if strings.EqualFold(token, command) {
-			return false
-		}
-	}
-	return true
-}
-
-func (p *commandProcessor) writeWorkflowCommandMessageLocked(target io.Writer, severity, message string) {
-	if message == "" {
-		return
-	}
-	p.writeMaskedLineLocked(target, severity+": "+message)
-}
-
-// trustedWarning records a runner-owned warning even after untrusted action
-// output has been suppressed, while preserving the job's registered masks.
-func (p *commandProcessor) trustedWarning(message string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.appendWorkflowCommandLocked(&p.trustedWarnings, workflowWarningAnnotationHeading, parsedWorkflowCommand{message: message})
-	p.writeWorkflowCommandMessageLocked(p.stderr, "warning", message)
-}
-
-func (p *commandProcessor) writeMaskedLineLocked(target io.Writer, line string) {
-	_, _ = fmt.Fprintln(target, p.maskTextLocked(line))
-}
-
-func (p *commandProcessor) writeLiteral(target io.Writer, line string) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	if !p.discard {
-		p.writeMaskedLineLocked(target, line)
-	}
-}
-
-func (p *commandProcessor) maskTextLocked(text string) string {
-	for _, mask := range p.masks {
-		text = strings.ReplaceAll(text, mask, "***")
-	}
-	return text
-}
-
-func (p *commandProcessor) appendWorkflowCommandLocked(buffer *workflowCommandAnnotationBuffer, heading string, command parsedWorkflowCommand) {
-	annotation := workflowCommandAnnotation{
-		file:     strings.Clone(commandText(command.properties["file"])),
-		title:    strings.Clone(commandText(command.properties["title"])),
-		location: strings.Clone(workflowCommandLocationLabel(command.properties)),
-		message:  strings.Clone(commandText(command.message)),
-	}
-	p.appendWorkflowCommandAnnotationLocked(buffer, heading, annotation)
-}
-
-func (p *commandProcessor) appendWorkflowCommandAnnotationLocked(buffer *workflowCommandAnnotationBuffer, heading string, command workflowCommandAnnotation) {
-	if buffer.truncated {
-		return
-	}
-	additional := len(renderWorkflowCommandListItem(command))
-	if len(buffer.commands) == 0 {
-		additional += len(heading) + len(workflowCommandListHeading) + len(workflowCommandListEnd)
-	}
-	if buffer.rendered+additional > maxJobAnnotationBytes-len(workflowCommandTruncationNotice) {
-		buffer.truncated = true
-		return
-	}
-	buffer.rendered += additional
-	buffer.commands = append(buffer.commands, command)
-}
-
-func (p *commandProcessor) suppress() {
-	p.mu.Lock()
-	p.discard = true
-	p.mu.Unlock()
-}
-
-func (p *commandProcessor) addMask(value string) {
-	p.mu.Lock()
-	p.addMaskLocked(value)
-	p.mu.Unlock()
-}
-
-func (p *commandProcessor) maskValues() []string {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return append([]string(nil), p.masks...)
-}
-
-type scrubbedCommandError struct {
-	cause   error
-	message string
-}
-
-func (e scrubbedCommandError) Error() string { return e.message }
-func (e scrubbedCommandError) Unwrap() error { return e.cause }
-
-func (p *commandProcessor) scrubError(err error) error {
-	if err == nil {
-		return nil
-	}
-	registered := p.maskValues()
-	masks := make([]string, 0, len(registered)*2)
-	for _, mask := range registered {
-		if mask == "" {
-			continue
-		}
-		masks = append(masks, mask)
-		quoted := strconv.Quote(mask)
-		escaped := quoted[1 : len(quoted)-1]
-		if escaped != mask {
-			masks = append(masks, escaped)
-		}
-	}
-	sort.Slice(masks, func(i, j int) bool {
-		if len(masks[i]) != len(masks[j]) {
-			return len(masks[i]) > len(masks[j])
-		}
-		return masks[i] < masks[j]
-	})
-	message := err.Error()
-	for _, mask := range masks {
-		message = strings.ReplaceAll(message, mask, "***")
-	}
-	if message == err.Error() {
-		return err
-	}
-	return scrubbedCommandError{cause: err, message: message}
-}
-
-func (p *commandProcessor) workflowCommandAnnotations() (warnings string, warningsTruncated bool, errors string, errorsTruncated bool) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	masks := normalizedMasks(p.masks)
-	var combinedWarnings workflowCommandAnnotationBuffer
-	for _, command := range p.trustedWarnings.commands {
-		p.appendWorkflowCommandAnnotationLocked(&combinedWarnings, workflowWarningAnnotationHeading, maskWorkflowCommandAnnotation(command, masks))
-	}
-	for _, command := range p.warnings.commands {
-		p.appendWorkflowCommandAnnotationLocked(&combinedWarnings, workflowWarningAnnotationHeading, maskWorkflowCommandAnnotation(command, masks))
-	}
-	combinedWarnings.truncated = combinedWarnings.truncated || p.trustedWarnings.truncated || p.warnings.truncated
-	warnings, warningsTruncated = renderWorkflowCommandAnnotation(workflowWarningAnnotationHeading, combinedWarnings.commands, combinedWarnings.truncated)
-	var maskedErrors workflowCommandAnnotationBuffer
-	for _, command := range p.errors.commands {
-		p.appendWorkflowCommandAnnotationLocked(&maskedErrors, workflowErrorAnnotationHeading, maskWorkflowCommandAnnotation(command, masks))
-	}
-	maskedErrors.truncated = maskedErrors.truncated || p.errors.truncated
-	errors, errorsTruncated = renderWorkflowCommandAnnotation(workflowErrorAnnotationHeading, maskedErrors.commands, maskedErrors.truncated)
-	return warnings, warningsTruncated, errors, errorsTruncated
-}
-
-func (p *commandProcessor) addMaskLocked(value string) {
-	if value == "" {
-		return
-	}
-	p.addMaskValueLocked(value)
-	for line := range strings.SplitSeq(strings.ReplaceAll(value, "\r\n", "\n"), "\n") {
-		if line != "" && line != value {
-			p.addMaskValueLocked(line)
-		}
-	}
-	sort.Slice(p.masks, func(i, j int) bool {
-		if len(p.masks[i]) != len(p.masks[j]) {
-			return len(p.masks[i]) > len(p.masks[j])
-		}
-		return p.masks[i] < p.masks[j]
-	})
-}
-
-func (p *commandProcessor) addMaskValueLocked(value string) {
-	if slices.Contains(p.masks, value) {
-		return
-	}
-	p.masks = append(p.masks, value)
 }
